app architecture visibility
