+++ demo/cmd/demo-wasm/nav.go
@@ -85,9 +85,12 @@ func (n *Nav) Render() string {
 func (n *Nav) OnMenuClick(s, e js.Value) {
 	app.NewContextMenu(
 		app.MenuItem{
-			Label:   "Reload",
-			Keys:    "cmdorctrl+r",
-			OnClick: app.Reload},
+			Label: "Reload",
+			Keys:  "cmdorctrl+r",
+			OnClick: func(s, e js.Value) {
+				app.Reload()
+			},
+		},
 		app.MenuItem{Separator: true},
 		app.MenuItem{
 			Label: "Go to repository",
